Add files via upload
The problem is I'm cteating a site and I neet to connect OpenServer to my site. So when I enter the data on site and press "sent"  I have to comeback to the main page of a site but it just show me this code I can get what I am supposted to change here. Please help me